refactor: fix circular dependency in redis service
